requirements.txt
fastapi==0.121.1
uvicorn[standard]==0.34.0
google-auth==2.41.1
google-auth-oauthlib==1.2.3
requests==2.32.5
requests-oauthlib==2.0.0
python-multipart==0.0.20
demail @ git+https://github.com/edire/demail.git@8649c5fc3ef47f85caca0fa51baeca4384b55d97
